SISRP-22288 - QA status and holds test fixes
